Update lwIP to 1.2.0

+++ gc/network.h
@@ -248,7 +248,7 @@ struct pollsd {
 };
 
 u32 inet_addr(const char *cp);
-s8 inet_aton(const char *cp, struct in_addr *addr);
+int inet_aton(const char *cp, struct in_addr *addr);
 char *inet_ntoa(struct in_addr addr); /* returns ptr to static buffer; not reentrant! */
 
 s32 if_config( char *local_ip, char *netmask, char *gateway,bool use_dhcp);
